// src/app/pages/calculator/models/calculator-settings.model.ts
export interface CalculatorSettings {
    nitrous_oxide: number
    methane: number
    synthetic_n_volatilization: number
    volatilization_n_residue: number
    urea_n_volatilization: number
    leaching: number
    soy_slope: number
    soy_intercept: number
    soy_nag: number
    soy_nbg: number
    soy_rbg_bio: number
    soy_frac_dm: number
    corn_slope_second_crop: number
    corn_intercept_second_crop: number
    corn_nag_second_crop: number
    corn_nbg_second_crop: number
    corn_rbg_bio_second_crop: number
    corn_frac_dm_second_crop: number
    diesel_b10: number
    regular_gasoline: number
    synthetic_fertilization: number
    urea_co2: number
    urea_n2o: number
    organic_n_fertilizer: number
    green_fertilizer_1: number
    green_fertilizer_2: number
    green_fertilizer_3: number
    atmospheric_deposition: number
    leaching_n_fertilizer: number
    calcitic_limestone: number
    dolomitic_limestone: number
    agricultural_plaster: number
    diesel_c02_mobile_source: number
    diesel_ch4_mobile_source: number
    diesel_n2o_mobile_source: number
    diesel_c02_stationary_source: number
    diesel_ch4_stationary_source: number
    diesel_n20_stationary_source: number
    consumption_gasoline: number
    organic_soils_cultivation: number
    nitrogen_loss: number
    preserved_native_forest_soil_1: number
    preserved_native_forest_soil_2: number
    preserved_native_forest_soil_3: number
    preserved_native_forest_soil_4: number
    preserved_native_forest_soil_5: number
    preserved_native_forest_soil_6: number
    preserved_native_forest_soil_7: number
    preserved_native_forest_soil_8: number
    preserved_native_forest_soil_9: number
    preserved_native_forest_soil_10: number
    preserved_native_forest_soil_11: number
    preserved_native_forest_soil_12: number
    preserved_native_forest_soil_13: number
    preserved_native_forest_soil_14: number
    preserved_native_forest_soil_15: number
    preserved_native_forest_soil_16: number
    preserved_native_forest_soil_17: number
    preserved_native_forest_soil_18: number
    preserved_native_forest_soil_19: number
    preserved_native_forest_soil_20: number
    preserved_native_forest_soil_21: number
    preserved_native_forest_soil_22: number
    preserved_native_forest_soil_23: number
    preserved_native_forest_soil_24: number
    preserved_native_forest_soil_25: number
    preserved_native_forest_soil_26: number
    preserved_native_forest_soil_27: number
    preserved_native_forest_soil_28: number
    preserved_native_forest_soil_29: number
    preserved_native_forest_soil_30: number
    preserved_native_forest_soil_31: number
    purchase_of_electricity_2019: number
    purchase_of_electricity_2020: number
    consumption_biodiesel_co2_mobile_source: number
    consumption_biodiesel_ch4_mobile_source: number
    consumption_biodiesel_n20_mobile_source: number
    consumption_ethanol_co2_stationary_source: number
    consumption_ethanol_ch4_stationary_source: number
    consumption_ethanol_n20_stationary_source: number
    nitrogen_fertilization_green_fertilizer: number
    land_occupation_1: number
    land_occupation_2: number
    land_occupation_3: number
    land_occupation_4: number
    land_occupation_5: number
    land_occupation_6: number
    land_occupation_7: number
    land_occupation_8: number
    land_occupation_9: number
    land_occupation_10: number
    land_occupation_11: number
    land_occupation_12: number
    land_occupation_13: number
    land_occupation_14: number
    land_occupation_15: number
    land_occupation_16: number
    land_occupation_17: number
    no_tillage_system_coverage: number
    no_tillage_system: number
    conventional_planting_system: number
}